Updated MakeReservationUserServlet
The logic of saving the reservation has been changed. Every time the user makes a reservation, he receives 10 points if he has less than 50, then the page is redirected to another jsp page that issues an alert about a successful reservation and prints a message that the user has received 10 points. If the user has 50 points then a discount of 10% is automatically assigned to him and a message is printed via alert that he has received a discount of 10%.