Created AtlantisAddTypeOfRoomServlet
A servlet that adds a new room type to a specific hotel. The hotel can have 2 types of rooms. So for now we have 1 input where we put the new "room type", after we press enter in the database it renames itself to the new room type we put in the input.