Created AtlantisParadiseBahamasHotelManagerPage.jsp
This is the main page of the manager. We have 1 hotel that can change, on that page there is a slide show with pictures of that 1 hotel (Atlantis Paradise Bahamas Hotel). Then we have a backend that connects to the database and from the database values ​​we put them into the hotel class values. So, after that, we write the information on the page in the table. For now, we have 3 inputs. The first input is for adding a new number of rooms, the second is for removing rooms and the third is for adding a new room type.