Changed a logic of editing text
changed logic for editing the text on the manager's side. Now we have an input field that, when filled, updates the text in the database and in the ConnectionDataBase class/model we have a new function that retrieves the hotelText. We also have a servlet on the AtlantisBahamasHotelManagerPage side that updates the field in the hotelText database. We are on the client side instantiated the ConnectionDataBase model and just called our new function that textHotel from the database.